+++ idoit-install.sh
@@ -224,6 +224,12 @@ function identifyOS {
         if [[ "$os_release" = "7.3" ]]; then
             log "Operating system identified as CentOS ${os_release}"
             OS="centos73"
+        elif [[ "$os_release" = "7.4" ]]; then
+            log "Operating system identified as CentOS ${os_release}"
+            OS="centos74"
+        elif [[ "$os_release" = "7.5" ]]; then
+            log "Operating system identified as CentOS ${os_release}"
+            OS="centos75"
         else
             abort "Operating system CentOS $os_release is not supported"
         fi
@@ -392,7 +398,7 @@ function configureOS {
         "ubuntu1604"|"ubuntu1610"|"ubuntu1704")
             configureUbuntu1604
             ;;
-        "rhel73"|"rhel74"|"rhel75"|"centos73")
+        "rhel73"|"rhel74"|"rhel75"|"centos73"|"centos74"|"centos75")
             configureRHEL
             ;;
         "sles12sp2"|"sles12sp3")
@@ -487,6 +493,16 @@ function configureRHEL {
         "centos73")
             os_description="CentOS"
             os_release="7.3"
+            mariadb_url="http://yum.mariadb.org/10.1/centos7-amd64"
+            ;;
+        "centos74")
+            os_description="CentOS"
+            os_release="7.4"
+            mariadb_url="http://yum.mariadb.org/10.1/centos7-amd64"
+            ;;
+        "centos75")
+            os_description="CentOS"
+            os_release="7.5"
             mariadb_url="http://yum.mariadb.org/10.1/centos7-amd64"
             ;;
     esac
@@ -674,7 +690,9 @@ function configurePHP {
 
     php_version=$(php --version | head -n1 -c7 | tail -c3)
 
-    if [[ "$OS" = "rhel73" || "$OS" = "rhel74" || "$OS" = "rhel75" || "$OS" = "centos73" ]]; then
+    if [[ "$OS" = "rhel73" || "$OS" = "rhel74" || "$OS" = "rhel75" ]]; then
+        ini_file="/etc/php.d/i-doit.ini"
+    elif [[ "$OS" = "centos73" || "$OS" = "centos74" || "$OS" = "centos75" ]]; then
         ini_file="/etc/php.d/i-doit.ini"
     elif [[ "$OS" = "sles12sp2" || "$OS" = "sles12sp3" ]]; then
         ini_file="/etc/php7/conf.d/i-doit.ini"
@@ -721,7 +739,7 @@ EOF
 
     log "Append path to MariaDB UNIX socket to PHP settings"
     case "$OS" in
-        "rhel73"|"rhel74"|"rhel75"|"centos73")
+        "rhel73"|"rhel74"|"rhel75"|"centos73"|"centos74"|"centos75")
             echo "mysqli.default_socket = /var/lib/mysql/mysql.sock" >> "$ini_file" || \
                 abort "Unable to alter PHP settings"
             ;;
@@ -751,7 +769,7 @@ function configureApache {
     log "Configure Apache Web server"
 
     case "$OS" in
-        "rhel73"|"rhel74"|"rhel75"|"centos73")
+        "rhel73"|"rhel74"|"rhel75"|"centos73"|"centos74"|"centos75")
             cat << EOF > /etc/httpd/conf.d/i-doit.conf || \
                 abort "Unable to create and edit file '/etc/httpd/conf.d/i-doit.conf'"
 <Directory /var/www/html/>
@@ -865,7 +883,7 @@ function configureMariaDB {
 
             mariadb_config="/etc/mysql/mariadb.conf.d/99-i-doit.cnf"
             ;;
-        "rhel73"|"rhel74"|"rhel75"|"centos73")
+        "rhel73"|"rhel74"|"rhel75"|"centos73"|"centos74"|"centos75")
             secureMariaDB
 
             mariadb_config="/etc/my.cnf.d/99-i-doit.cnf"