+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -   Support for Red Hat Enterprise Linux (RHEL) 7.5
 -   Match available CPU cores and RAM with requirements
+-   `idoit-hotfix`: Deploy hot fixes
 
 
 ### Fixed